Updated to 16.0.6 Added prep section to spec file to confirm tar.bz2 file integrity Minor updates to README file

+++ centos/nextcloud.spec
@@ -13,12 +13,17 @@
 
 Summary: Nextcloud package
 Name: nextcloud
-Version: 16.0.4
+Version: 16.0.6
 Release: 1%{?dist}
 License: GPL
 Source: https://download.nextcloud.com/server/releases/nextcloud-%{version}.tar.bz2
-Source1: nextcloud.conf
-Source2: nextcloud-fpm.conf
+Source1: https://raw.githubusercontent.com/nextcloud/server-packages/master/centos/nextcloud.conf
+Source2: https://raw.githubusercontent.com/nextcloud/server-packages/master/centos/nextcloud-fpm.conf
+Source3: https://nextcloud.com/nextcloud.asc
+Source4: https://download.nextcloud.com/server/releases/nextcloud-%{version}.tar.bz2.asc
+Source5: https://download.nextcloud.com/server/releases/nextcloud-%{version}.tar.bz2.md5
+Source6: https://download.nextcloud.com/server/releases/nextcloud-%{version}.tar.bz2.sha256
+Source7: https://download.nextcloud.com/server/releases/nextcloud-%{version}.tar.bz2.sha512
 BuildArch: noarch
 URL: https://nextcloud.com/
 
@@ -42,6 +47,8 @@ Requires: rh-php72-php-ldap
 # Required php packages for MariaDB
 Requires: rh-php72-php-pdo_mysql
 
+# NextCloud does not support skipping a major version number
+Conflicts: nextcloud < 15
 
 %description
 Nextcloud files and configuration.
@@ -51,6 +58,17 @@ nc_dir:        %{nc_dir}
 nc_data_dir:   %{nc_data_dir}
 nc_config_dir: %{nc_config_dir}
 
+%prep
+cd %{_sourcedir}
+/usr/bin/md5sum -c %{SOURCE5}
+if [ $? -ne 0 ] ; then echo md5sum did not match ; exit 1 ; fi
+/usr/bin/sha256sum -c %{SOURCE6}
+if [ $? -ne 0 ] ; then echo sha256sum did not match ; exit 1 ; fi
+/usr/bin/sha512sum -c %{SOURCE7}
+if [ $? -ne 0 ] ; then echo sha512sum did not match ; exit 1 ; fi
+/usr/bin/gpg --import %{SOURCE3}
+/usr/bin/gpg --verify %{SOURCE4} %{SOURCE0}
+if [ $? -ne 0 ] ; then echo gpg signature did not match ; exit 1 ; fi
 
 %install
 rm -rf %{buildroot}
@@ -103,6 +121,10 @@ cp %{SOURCE2} %{buildroot}/etc/opt/rh/rh-php72/php-fpm.d/nextcloud.conf
 
 
 %changelog
+* Thu Nov 7 2019 B Galliart <ben@steadfast.net> - 16.0.6-1
+- Update to release 16.0.6
+- Added prep section to verify integrity of nextcloud tar.bz2 file
+
 * Thu Sep 02 2019 Giacomo Sanchietti <giacomo.sanchietti@nethesis.it>- 16.0.4-1
 - Update to release 16.0.4
